re-organized application layer
