+++ src/main/java/es/thalesalv/chatrpg/application/commands/dmassist/PromptDMAssistCommand.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 
 import es.thalesalv.chatrpg.adapters.data.db.repository.ChannelRepository;
 import es.thalesalv.chatrpg.application.commands.DiscordCommand;
-import es.thalesalv.chatrpg.application.service.GptModelService;
+import es.thalesalv.chatrpg.application.service.completion.TextCompletionService;
 import es.thalesalv.chatrpg.application.service.usecases.BotUseCase;
 import es.thalesalv.chatrpg.application.translator.ChannelEntityListToDTOList;
 import es.thalesalv.chatrpg.application.translator.MessageEventDataTranslator;
@@ -93,7 +93,7 @@ public class PromptDMAssistCommand extends DiscordCommand {
                 final MessageEventData messageEventData = messageEventDataTranslator.translate(event.getJDA()
                         .getSelfUser(), channel, contextDatastore.getCommandEventData().getChannelConfig(), message);
 
-                final GptModelService model = (GptModelService) applicationContext.getBean(modelSettings.getModelFamily() + MODEL_SERVICE);
+                final TextCompletionService model = (TextCompletionService) applicationContext.getBean(modelSettings.getModelFamily() + MODEL_SERVICE);
                 final BotUseCase useCase = (BotUseCase) applicationContext.getBean(persona.getIntent() + USE_CASE);
 
                 useCase.generateResponse(messageEventData, model);
