+++ src/main/java/es/thalesalv/chatrpg/application/service/completion/TextCompletionService.java
@@ -1,11 +1,11 @@
-package es.thalesalv.chatrpg.application.service;
+package es.thalesalv.chatrpg.application.service.completion;
 
 import java.util.List;
 
 import es.thalesalv.chatrpg.domain.model.openai.dto.MessageEventData;
 import reactor.core.publisher.Mono;
 
-public interface GptModelService {
+public interface TextCompletionService {
 
     Mono<String> generate(final String prompt, final List<String> messages, final MessageEventData messageEventData);
 }
